second_step

+++ backend/controllers/certificateController.js
@@ -1,28 +1,26 @@
 import Certificate from '../models/Certificate.js'
+import Result from '../models/Result.js'
 
 const addCertificate = async (req, res) => {
   try {
-    const { Loai, NgonNgu, TenChungChi, CapDo, LePhiThi } = req.body
+    const { Loai, TenChungChi, LePhiThi, ThoiHan } = req.body
 
-    const newCertificate = new Certificate({ Loai, NgonNgu, TenChungChi, CapDo, LePhiThi })
+    const newCertificate = new Certificate({ Loai, TenChungChi, LePhiThi, ThoiHan })
     await newCertificate.save()
 
-    res.status(201).json({ message: 'Thêm chứng chỉ thành công' })
+    res.status(201).json({ message: 'Thêm chứng chỉ thành công', data: newCertificate })
   } catch (error) {
-    res.status(500).json({ error: error.message })
+    res.status(500).json({ message: 'Lỗi server', error: error.message })
   }
 }
 
 const getCertificates = async (req, res) => {
   try {
     const certificates = await Certificate.find()
-    if (!certificates) {
-      return res.status(404).json({ message: 'Không tìm thấy chứng chỉ nào'})
-    }
 
     res.status(200).json(certificates)
   } catch (error) {
-    res.status(500).json({ error: error.message }) 
+    res.status(500).json({ message: 'Lỗi server', error: error.message })
   }
 }
 
@@ -32,34 +30,34 @@ const updateCertificate = async (req, res) => {
     const updates = req.body
 
     const updatedCertificate = await Certificate.findByIdAndUpdate(
-      id,
+      { id }, 
       { $set: updates },
       { new: true }
-    ) 
-
-    if (!updatedCertificate) {
-      return res.status(404).json({ message: 'Không tìm thấy chứng chỉ để cập nhật' })
-    }
+    )
 
     res.status(200).json(updatedCertificate)
   } catch (error) {
-    res.status(500).json({ error: error.message }) 
+    res.status(500).json({ message: 'Lỗi server', error: error.message })
   }
-}
+};
 
 const deleteCertificate = async (req, res) => {
   try {
-    const { id } = req.params 
+    const { id } = req.params
 
-    const deletedCertificate = await Certificate.findByIdAndDelete(id) 
-
-    if (!deletedCertificate) {
-      return res.status(404).json({ message: 'Không tìm thấy chứng chỉ để xóa' }) 
+    const referencedResult = await Result.findOne({ IDChungChi: id })
+    if (referencedResult) {
+      return res.status(400).json({ message: 'Không thể xóa' })
     }
 
-    res.status(200).json({ message: 'Xóa chứng chỉ thành công' }) 
+    const deletedCertificate = await Certificate.findByIdAndDelete(id)
+    if (!deletedCertificate) {
+      return res.status(404).json({ message: 'Không tìm thấy chứng chỉ để xóa' })
+    }
+
+    res.status(200).json({ message: 'Xóa chứng chỉ thành công' })
   } catch (error) {
-    res.status(500).json({ error: error.message }) 
+    res.status(500).json({ message: 'Lỗi server', error: error.message })
   }
 }
 
@@ -67,5 +65,5 @@ export default {
   addCertificate,
   getCertificates,
   updateCertificate,
-  deleteCertificate
+  deleteCertificate 
 }
